Declare buffers as float in Exercise 5

+++ Code_Exercises/Data_and_Dependencies/solution.cpp
@@ -95,7 +95,7 @@ void test_buffer() {
 void test_usm() {
   constexpr size_t dataSize = 1024;
 
-  int inA[dataSize], inB[dataSize], inC[dataSize], out[dataSize];
+  float inA[dataSize], inB[dataSize], inC[dataSize], out[dataSize];
   for (int i = 0; i < dataSize; ++i) {
     inA[i] = static_cast<float>(i);
     inB[i] = static_cast<float>(i);
